+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * [UIBULKED-653](https://folio-org.atlassian.net/browse/UIBULKED-653) Add "Create profile" form
 * [UIBULKED-663](https://folio-org.atlassian.net/browse/UIBULKED-663) Deleting a bulk edit profile
 * [UIBULKED-658](https://folio-org.atlassian.net/browse/UIBULKED-658) Add "Update profile" form
+* [UIBULKED-674](https://folio-org.atlassian.net/browse/UIBULKED-674) View "Bulk edits" of existing bulk edit profile
 
 ## [5.0.0](https://github.com/folio-org/ui-bulk-edit/tree/v5.0.0) (2025-03-12)